CSP-501 Removed Identity setting from Profile table

+++ src/SFA.DAS.AANHub.Database/Tables/MemberProfile.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[MemberProfile]
 (
-	[MemberId] UNIQUEIDENTIFIER NOT NULL PRIMARY KEY, 
-    [ProfileId] BIGINT NOT NULL, 
+    [MemberId] UNIQUEIDENTIFIER NOT NULL PRIMARY KEY, 
+    [ProfileId] INT NOT NULL, 
     [ProfileValue] NVARCHAR(200) NOT NULL
     CONSTRAINT [FK_MemberProfile_Member] FOREIGN KEY ([MemberId]) REFERENCES [Member]([Id])
     CONSTRAINT [FK_MemberProfile_Profile] FOREIGN KEY ([ProfileId]) REFERENCES [Profile]([Id])
